Version the state properly and stop deleting records from the client's cache
The current version of bicycle breaks in poor network conditions because the server assumes that any response it sends is guaranteed to be delivered to the client.  This can lead to a permanent loading state where:

1. the client asks for some data
2. the server attempts to send that data, updating it's record of what data the client currently has.
3. the client fails to receive the response, so it re-requests the data from the server.
4. the server checks what it thinks the client already has, and doesn't send any new data, even though the client has not yet received anything.

By versioning the store on the server, we can ensure that the client will re-request with the *old* version of the state on the server.  The server can then see that the client is missing data and send it as needed.

I've (separately) removed the ability for the server to delete data from the client. This allows for the client to instantly load data (albeit potentially stale) whenever that data has already been loaded before, even if it is not in the current query set.  Since the client side cache is normalised, this should still represent a good user experience.